requirements.txt
nltk
pymorphy3
pytest
python-telegram-bot
tqdm